가방조회(attackInfo)에 GP 잔액 표시 추가 (#95)
- 포인트 바로 아래 'GP: X.XX' 한 줄 추가 (소수점 2자리)
- selectGpBalance 반환 타입 int→double로 수정 (0.20, 0.85 등 소수 GP 정확히 집계)

+++ src/main/java/my/prac/core/prjbot/dao/BotNewDAO.java
@@ -203,7 +203,7 @@ public interface BotNewDAO {
     List<Integer> selectItemIdsByType(@Param("itemType") String itemType);
     List<HashMap<String, Object>> selectHellRewardItemsWithOwnCount();
     void insertGpRecord(HashMap<String, Object> param);
-    int selectGpBalance(@Param("userName") String userName);
+    double selectGpBalance(@Param("userName") String userName);
 
     void lockMacroUser(HashMap<String, Object> map);
